Database review: standardise schema, optimise queries, tighten column types
Standardise all FK constraints to explicit ON DELETE RESTRICT, unify string
columns on TEXT, narrow oversized INT columns to SMALLINT, replace full-table
DISTINCT ON / ROW_NUMBER subqueries with scoped LATERAL joins, rewrite
non-sargable date casts to range predicates, add partial index for current
club members, remove dead code, add retention methods, and reorder companion
object methods to a consistent convention.

+++ src/main/scala/ccas/analysis/apps/history/HistoryProcessing.scala
@@ -194,11 +194,11 @@ private[history] object HistoryProcessing {
         val team1Fp = teams.team1.fairPlayRemovals.map(_.value)
         val team2Fp = teams.team2.fairPlayRemovals.map(_.value)
 
-        val team1ByBoard: Map[Int, MatchPlayerStarted] = teams.team1.players.collect {
-          case p: MatchPlayerStarted => p.board.path.segments.last.toInt -> p
+        val team1ByBoard: Map[Short, MatchPlayerStarted] = teams.team1.players.collect {
+          case p: MatchPlayerStarted => p.board.path.segments.last.toShort -> p
         }.toMap
-        val team2ByBoard: Map[Int, MatchPlayerStarted] = teams.team2.players.collect {
-          case p: MatchPlayerStarted => p.board.path.segments.last.toInt -> p
+        val team2ByBoard: Map[Short, MatchPlayerStarted] = teams.team2.players.collect {
+          case p: MatchPlayerStarted => p.board.path.segments.last.toShort -> p
         }.toMap
 
         val allBoards = (team1ByBoard.keySet ++ team2ByBoard.keySet).toList.sorted
@@ -518,11 +518,11 @@ private[history] object HistoryProcessing {
       timeClass = dailyMatch.settings.timeClass,
       startTime = startTime,
       endTime = endTime,
-      boards = dailyMatch.boards,
+      boards = dailyMatch.boards.toShort,
       team1ClubId = team1ClubId,
-      team1ScoreX2 = (teams.team1.score * 2).toInt,
+      team1ScoreX2 = (teams.team1.score * 2).toShort,
       team2ClubId = team2ClubId,
-      team2ScoreX2 = (teams.team2.score * 2).toInt,
+      team2ScoreX2 = (teams.team2.score * 2).toShort,
       fetchedAt = Instant.now()
     )
   }
